DualOdometryCorrection, and Demo added
DualOdometryCorrection was added with a method to allow the thread to
correct the angle or not so that we don't always do it and end up with a
slow robot.

MasterBrick was updated, there is no colorsensor as it is probably
better and simpler to do it in a different thread. The getUSFilteredData
replaces getFilteredData as it doesn't need to be universal anymore, but
only to correspond to the UltrasonicSensor.
Odometer was updated with setX, setY and setAng added.
Demo contains the main method for testing.